+++ ui/v2.5/src/components/Tagger/scenes/SceneTagger.tsx
@@ -7,15 +7,15 @@ import { FormattedMessage, useIntl } from "react-intl";
 import { Icon } from "src/components/Shared/Icon";
 import { LoadingIndicator } from "src/components/Shared/LoadingIndicator";
 import { OperationButton } from "src/components/Shared/OperationButton";
-import { IScrapedScene, TaggerStateContext } from "../context";
+import { TaggerStateContext } from "../context";
 import Config from "./Config";
 import { TaggerScene } from "./TaggerScene";
 import { SceneTaggerModals } from "./sceneTaggerModals";
 import { SceneSearchResults } from "./StashSearchResult";
 import { ConfigurationContext } from "src/hooks/Config";
 import { faCog } from "@fortawesome/free-solid-svg-icons";
-import { distance } from "src/utils/hamming";
 import { useLightbox } from "src/hooks/Lightbox/hooks";
+import { compareScenesForSort } from "./utils";
 
 interface ITaggerProps {
   scenes: GQL.SlimSceneDataFragment[];
@@ -91,139 +91,6 @@ export const Tagger: React.FC<ITaggerProps> = ({ scenes, queue }) => {
         </Button>
       </div>
     );
-  }
-
-  function minDistance(hash: string, stashScene: GQL.SlimSceneDataFragment) {
-    let ret = 9999;
-    stashScene.files.forEach((cv) => {
-      if (ret === 0) return;
-
-      const stashHash = cv.fingerprints.find((fp) => fp.type === "phash");
-      if (!stashHash) {
-        return;
-      }
-
-      const d = distance(hash, stashHash.value);
-      if (d < ret) {
-        ret = d;
-      }
-    });
-
-    return ret;
-  }
-
-  function calculatePhashComparisonScore(
-    stashScene: GQL.SlimSceneDataFragment,
-    scrapedScene: IScrapedScene
-  ) {
-    const phashFingerprints =
-      scrapedScene.fingerprints?.filter((f) => f.algorithm === "PHASH") ?? [];
-    const filteredFingerprints = phashFingerprints.filter(
-      (f) => minDistance(f.hash, stashScene) <= 8
-    );
-
-    if (phashFingerprints.length == 0) return [0, 0];
-
-    return [
-      filteredFingerprints.length,
-      filteredFingerprints.length / phashFingerprints.length,
-    ];
-  }
-
-  function minDurationDiff(
-    stashScene: GQL.SlimSceneDataFragment,
-    duration: number
-  ) {
-    let ret = 9999;
-    stashScene.files.forEach((cv) => {
-      if (ret === 0) return;
-
-      const d = Math.abs(duration - cv.duration);
-      if (d < ret) {
-        ret = d;
-      }
-    });
-
-    return ret;
-  }
-
-  function calculateDurationComparisonScore(
-    stashScene: GQL.SlimSceneDataFragment,
-    scrapedScene: IScrapedScene
-  ) {
-    if (scrapedScene.fingerprints && scrapedScene.fingerprints.length > 0) {
-      const durations = scrapedScene.fingerprints.map((f) => f.duration);
-      const diffs = durations.map((d) => minDurationDiff(stashScene, d));
-      const filteredDurations = diffs.filter((duration) => duration <= 5);
-
-      const minDiff = Math.min(...diffs);
-
-      return [
-        filteredDurations.length,
-        filteredDurations.length / durations.length,
-        minDiff,
-      ];
-    }
-    return [0, 0, 0];
-  }
-
-  function compareScenesForSort(
-    stashScene: GQL.SlimSceneDataFragment,
-    sceneA: IScrapedScene,
-    sceneB: IScrapedScene
-  ) {
-    // Compare sceneA and sceneB to each other for sorting based on similarity to stashScene
-    // Order of priority is: nb. phash match > nb. duration match > ratio duration match > ratio phash match
-
-    // scenes without any fingerprints should be sorted to the end
-    if (!sceneA.fingerprints?.length && sceneB.fingerprints?.length) {
-      return 1;
-    }
-    if (!sceneB.fingerprints?.length && sceneA.fingerprints?.length) {
-      return -1;
-    }
-
-    const [nbPhashMatchSceneA, ratioPhashMatchSceneA] =
-      calculatePhashComparisonScore(stashScene, sceneA);
-    const [nbPhashMatchSceneB, ratioPhashMatchSceneB] =
-      calculatePhashComparisonScore(stashScene, sceneB);
-
-    // If only one scene has matching phash, prefer that scene
-    if (
-      (nbPhashMatchSceneA != nbPhashMatchSceneB && nbPhashMatchSceneA === 0) ||
-      nbPhashMatchSceneB === 0
-    ) {
-      return nbPhashMatchSceneB - nbPhashMatchSceneA;
-    }
-
-    // Prefer scene with highest ratio of phash matches
-    if (ratioPhashMatchSceneA !== ratioPhashMatchSceneB) {
-      return ratioPhashMatchSceneB - ratioPhashMatchSceneA;
-    }
-
-    // Same ratio of phash matches, check duration
-    const [
-      nbDurationMatchSceneA,
-      ratioDurationMatchSceneA,
-      minDurationDiffSceneA,
-    ] = calculateDurationComparisonScore(stashScene, sceneA);
-    const [
-      nbDurationMatchSceneB,
-      ratioDurationMatchSceneB,
-      minDurationDiffSceneB,
-    ] = calculateDurationComparisonScore(stashScene, sceneB);
-
-    if (nbDurationMatchSceneA != nbDurationMatchSceneB) {
-      return nbDurationMatchSceneB - nbDurationMatchSceneA;
-    }
-
-    // Same number of phash & duration, check duration ratio
-    if (ratioDurationMatchSceneA != ratioDurationMatchSceneB) {
-      return ratioDurationMatchSceneB - ratioDurationMatchSceneA;
-    }
-
-    // fall back to duration difference - less is better
-    return minDurationDiffSceneA - minDurationDiffSceneB;
   }
 
   const [spriteImage, setSpriteImage] = useState<string | null>(null);
